Improve CI job for running tests
Adapted the execution of cargo nextest to use a CI specific profile and
produce test results in JUnit format.
Included the running of the doctests in the same job that runs cargo
nextest so that the doctests are also executed for all relevant
feature combinations.
Merged the test results into a single archive so that all results are
part of the archive that is attached to releases.