[#50] Conditionally build tests requiring Docker
Some of the integration tests require Docker to start a Mosquitto
broker. These tests fail if Docker is not available on the build
platform.

A build.rs script has been added which checks if Docker is available
and allows tests to be conditionally built (and run) if that is indeed
the case.