Add integration tests verifying reconnect behavior of the MQTT client
Adapted check CI job to use standard cargo test instead of nextest
because the latter does not support the test-serial crate which is
needed for running the integration tests sequentially so that the
Mosquitto container can be run with a fixed host port. This is
necessary because the container is being stopped and restarted as
part of the test scenarios.